Removed filterBookingsWithoutReaction from IncomingReactionModalContainer
